requirements.txt
ibm_watson_machine_learning
langchain
langchain-ibm
numpy==1.26.1
openpyxl==3.1.2
pandas==1.5.3
python-dotenv==1.0.0
XlsxWriter==3.2.0
